Remove CAA lookup support pending more thought/design.
We're not presently interested in looking at CAA testing with this tool
and the semantics of tree-climbing make it a little bit difficult to
integrate into the codebase as-is. Rip out CAA support for now, we can
always add it back when we have more time and a specific use-case in
mind.