Big refactor, better multi-server support, DB container for dev. (#2)
This is a big PR that refactors the old `caa-lookups` based code initially
imported into this repo. It reorganizes things slightly by pulling most of the
code out of the `dnslol` command into a `dnslol` package that
`cmd/dnslol/main.go` can reference. In the future this will make writing unit
tests easier.

The changes in this branch lay the ground work for comparing results between
nameservers. The old `caa-lookups` code would only round-robin between
configured servers. Now it is assumed queries should be sent to all configured
servers for comparison. The output & Prometheus stats are updated to ensure its
possible to differentiate results between servers.

In preparation for storing the results in a database instead of printing them to
standard out for processing with shell based tools this branch includes a small
`docker-compose.yml` file that can be used to stand up a MariaDB instance. Right
now nothing is using this DB instance but that's where I'm heading next
development-wise.