Initial database support for persisting results.
Three database tables are specified in the provided `db-schema.sql`:
1) Experiments - holding an ID for each experiment, the start time,the
   end time, and the `dnslol` command line arguments used.
2) Servers - holding an ID for each individual DNS server address, an ID
   for the corresponding Experiment, and the hostname of each DNS
   server. This table allows an experiment to have many servers.
3) Results - holding an ID for each individual query result, the name
   that was queried, the type of query that was performed, the recorded
   error (if any), the ID of the server that was queried, and the ID of
   the experiment that the query was associated with.

`dnslol` is updated to accept a MySQL DSN to specify the
username/password/host/databasename for the DB connection. When an
Experiment is started the corresponding Experiments and Servers rows are
created. Results are written to the Results table as they are received.