Initial database support for persisting results. (#3)
Three database tables are specified in the provided `db-schema.sql`:
1) Experiments - holding an ID for each experiment, the start time, the end time, and the `dnslol` command line arguments used.
2) Servers - holding an ID for each individual DNS server address, an ID for the corresponding Experiment, and the hostname of each DNS server. This table allows an experiment to have many servers.
3) Results - holding an ID for each individual query result, the name that was queried, the type of query that was performed, the recorded error (if any), the ID of the server that was queried, and the ID of the experiment that the query was associated with.

`dnslol` is updated to accept a MySQL DSN to specify the username/password/host/databasename for the DB connection. When an Experiment is started the corresponding Experiments and Servers rows are created. Results are written to the Results table as they are received.

A new `-count` argument is added to `dnslol` to allow doing `count` copies of each query per nameserver within one experiment. This will make it easier to find flaky names.

After discovering some unicode IDN domains snuck into the dataset (causing DB persist errors due to MySQL `utf8` vs `utf8mb4`) this branch also updates `dnslol` to reject non-ascii domains. The README is updated to describe this and to provide a small Go program that can convert IDN to IDNA2008 encoded ascii.